requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
black==25.11.0
bleach==6.2.0
blinker==1.9.0
cachetools==6.1.0
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
datasets==4.4.1
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
distlib==0.3.8
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
executing==2.2.0
fastapi==0.124.0
fastapi-cli==0.0.16
fastapi-cloud-cli==0.6.0
fastar==0.8.0
fastjsonschema==2.21.1
filelock==3.13.1
fonttools==4.59.0
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.44
groq==1.1.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.10
imbalanced-learn==0.14.0
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kaggle==1.8.0
kagglesdk==0.1.13
kiwisolver==1.4.8
lark==1.2.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
mypy==1.18.2
mypy_extensions==1.1.0
narwhals==1.48.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
notebook==7.4.4
notebook_shim==0.2.4
numpy==2.3.1
orjson==3.11.5
overrides==7.7.0
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
patsy==1.0.2
pillow==11.3.0
pipenv==2023.12.1
platformdirs==4.2.0
plotly==6.4.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.4.1
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.22
pydantic==2.12.5
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==3.3.0
python-multipart==0.0.20
python-slugify==8.0.4
pytokens==0.3.0
pytz==2025.2
pywin32==311
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.10
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
rpds-py==0.26.0
safetensors==0.7.0
scikit-learn==1.7.1
scipy==1.16.0
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.47.0
setuptools==69.2.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.50.0
statsmodels==0.14.5
streamlit==1.47.0
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.1+cpu
torchaudio==2.9.1+cpu
torchvision==0.24.1+cpu
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
typer==0.20.0
types-python-dateutil==2.9.0.20250708
types-requests==2.32.4.20250913
types-tqdm==4.67.0.20250809
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.11.0
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.38.0
virtualenv==20.25.1
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
